pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-phonemask-additions"
version = "1.2.0"
dependencies = [
    "django>=4.2",
    "requests>=2.25.1",
    "phonenumbers>=8.12.0",
    "psycopg2-binary>=2.9.10",
    "python-docx"
]
description = "A reusable Django app for user authentication, phone validation, and admin management."
readme = "README.rst"
requires-python = ">=3.10"
authors = [{name = "Lag Bag", email = "vitya-dolzhenkov@mail.ru"}]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

[project.urls]
Homepage = "https://github.com/LagBag/django-phonemask-additions"